Revert "Revert "Revert "DG-1869 Evaluates API Optimisation"""
